// libs/model/convs-mgr/stories/flows/src/lib/editor/flow-control-type.enum.ts
export enum FlowControlType
{
  Header = 'h1',
  LightHeader = 'h2',
  Text = 'text',
  Caption = 'caption',
  Image = 'image',
  Link = 'link',
  Footer = 'footer',
  TextInput = 'textInput',
  TextArea = 'textArea',
  Select = 'select',
  Radio = 'radio',
  OptIn = 'checkbox',
  Datepick = 'datepick'
}